+++ docs/settings.py
@@ -71,4 +71,5 @@ class Site(Site,ledger.SiteMixin):
         #~ yield 'lino'
 
 SITE = Site(globals())
+print 20130409, __file__, LOGGING