+++ lino/projects/cosi/tests/__init__.py
@@ -1,0 +1,1 @@
+from cosi_tests import *