+++ lino/ui/tests/lino_ui_tests.py
@@ -33,14 +33,10 @@ from django.utils.encoding import force_unicode
 from django.core.exceptions import ValidationError
 
 #~ from lino import dd
-#~ from lino.utils import i2d
-#~ from lino.core.dbutils import resolve_model
-#Companies = resolve_model('contacts.Companies')
-from djangosite.utils.test import TestCase
+from djangosite.utils.test import NoAuthTestCase
 
 
-
-class QuickTest(TestCase):
+class QuickTest(NoAuthTestCase):
     pass